+++ src/plugins/weather_station/ws_config.cpp
@@ -14,11 +14,14 @@ namespace KFCConfigurationClasses {
         cfg.additionalClocks[0]._enabled = true;
         cfg.additionalClocks[1]._enabled = true;
         cfg.additionalClocks[2]._enabled = true;
+        // New York, USA
+        cfg.latitude = 40.743498405083564;
+        cfg.longitude = -74.0285408439419;
         setConfig(cfg);
         setTZ0(F("Berlin" "\xff" "CET-1CEST,M3.5.0,M10.5.0/3" "\xff" "Europe/Berlin"));
         setTZ1(F("New York" "\xff" "EST5EDT,M3.2.0,M11.1.0" "\xff" "America/New_York"));
         setTZ2(F("Tokyo" "\xff" "JST-9" "\xff" "Asia/Tokyo"));
         setApiKey(F("GET_YOUR_API_KEY_openweathermap.org"));
-        setApiQuery(F("New York,US"));
+        setLocation(F("New York,US"));
     }
 }